updated script

+++ src/UserLogin.js
@@ -10,23 +10,28 @@ export function UserLogin(){
 
     const [user_email, setUserEmail] = useState("");
     const [user_pass, setUserPass] = useState("");
+    const [confPass, setConfPass] = useState("");
     const navigate = useNavigate();
     
     function onUserLogin(event){
         event.preventDefault();
-        axios.post(liveUrl + "UserLogin", {user_email, user_pass})
-        .then(res => {
-            if(res.status !== 200){
-                //Display error
-            }else{
-                localStorage.clear();
-                localStorage.setItem("Token", res.data);
-                localStorage.setItem("ID", user_email);
-                userToken = res.data;
-                navigate("/UserStore");
-            }
-        })
-        .catch(err => console.log(err))
+        if(confPass!=user_pass){
+            window.alert("Password does not match!");
+        }else{
+            axios.post(liveUrl + "UserLogin", {user_email, user_pass})
+            .then(res => {
+                if(res.status !== 200){
+                    //Display error
+                }else{
+                    localStorage.clear();
+                    localStorage.setItem("Token", res.data);
+                    localStorage.setItem("ID", user_email);
+                    userToken = res.data;
+                    navigate("/UserStore");
+                }
+            })
+            .catch(err => console.log(err))
+        }
     }
 
     return <>
@@ -41,8 +46,8 @@ export function UserLogin(){
                 <input type="password" name="password" id="userPassword" className="formElement" onChange={ (e)=>{setUserPass(e.target.value)} }/>
 
                 <label className="formElement">Confirm your password :&emsp;</label>
-                <input className="formElement" />
-                    
+                <input className="formElement" type="password" onChange={(e)=>{setConfPass(e.target.value)}}/>
+
                 <button type="submit" className="formElement">Submit</button>
 
             </form>
@@ -56,16 +61,21 @@ export function UserSignup(){
     const [username, setUserName] = useState("");
     const [user_email, setUserEmail] = useState("");
     const [user_pass, setUserPass] = useState("");
+    const [confPass, setConfPass] = useState("");
 
     const navigate = useNavigate();
     
     function onUserSignup(event){
         event.preventDefault();
-        axios.post(liveUrl + "UserSignup", {user_pass, username, user_email})
-        .then(res => {console.log(res)
-        navigate("/UserLogin");
-        })
-        .catch(err => console.log(err))
+        if(confPass!=user_pass){
+            window.alert("Password does not match!");
+        }else{
+            axios.post(liveUrl + "UserSignup", {user_pass, username, user_email})
+            .then(res => {console.log(res)
+            navigate("/UserLogin");
+            })
+            .catch(err => console.log(err))
+        }
     }
 
     return <>
@@ -83,8 +93,8 @@ export function UserSignup(){
                 <input type="password" name="password" id="userPassword" className="formElement" onChange={ (e)=>{setUserPass(e.target.value)} }/>
 
                 <label className="formElement">Confirm your password :&emsp;</label>
-                <input className="formElement" />
-                    
+                <input className="formElement" type="password" onChange={(e)=>{setConfPass(e.target.value)}}/>
+
                 <button type="submit" className="formElement">Submit</button>
             </form>
             <p></p>
@@ -95,22 +105,28 @@ export function UserSignup(){
 export function SellerLogin(){
     const [seller_email, setSellerEmail] = useState("");
     const [seller_pass, setSellerPass] = useState("");
+    const [confPass, setConfPass] = useState("");
     const navigate = useNavigate();
     function onSellerLogin(event){
         localStorage.clear();
         event.preventDefault();
-        axios.post(liveUrl + "SellerLogin", {seller_email, seller_pass})
-        .then(res => {
-            if(res.status !== 200){
-                //Display error
-            }else{
-                localStorage.setItem("Token", res.data);
-                localStorage.setItem("ID", seller_email);
-                userToken = res.data;
-                navigate("/SellerStore");
-            }
-        })
-        .catch(err => {console.log(err)})
+        if(confPass!=seller_pass){
+            window.alert("Password does not match!");
+        }else{
+            axios.post(liveUrl + "SellerLogin", {seller_email, seller_pass})
+            .then(res => {
+                if(res.status !== 200){
+                    //Display error
+                }else{
+                    localStorage.setItem("Token", res.data);
+                    localStorage.setItem("ID", seller_email);
+                    userToken = res.data;
+                    navigate("/SellerStore");
+                }
+            })
+            .catch(err => {console.log(err)})
+        }
+        
     }
 
     return <>
@@ -125,8 +141,8 @@ export function SellerLogin(){
                 <input type="password" name="password" id="userPassword" className="formElement" onChange={ (e)=>{setSellerPass(e.target.value)} }/>
 
                 <label className="formElement">Confirm your password :&emsp;</label>
-                <input className="formElement"/>
-                    
+                <input className="formElement" type="password" onChange={(e)=>{setConfPass(e.target.value)}}/>
+                
                 <button type="submit" className="formElement">Submit</button>
             </form>
             <p></p>
@@ -139,16 +155,21 @@ export function SellerSignup(){
     const [seller_name, setSellerName] = useState("");
     const [seller_email, setSellerEmail] = useState("");
     const [seller_pass, setSellerPass] = useState("");
+    const [confPass, setConfPass] = useState("");
 
     const navigate = useNavigate();
     
     function onSellerSignup(event){
         event.preventDefault();
-        axios.post(liveUrl + "SellerSignup", {seller_pass, seller_name, seller_email})
-        .then(res => {console.log(res)
-        navigate("/SellerLogin");
-        })
-        .catch(err => console.log(err))
+        if(confPass!=seller_pass){
+            window.alert("Password does not match!")
+        }else{
+            axios.post(liveUrl + "SellerSignup", {seller_pass, seller_name, seller_email})
+            .then(res => {console.log(res)
+            navigate("/SellerLogin");
+            })
+            .catch(err => console.log(err))
+        }
     }
     
     return <>
@@ -169,7 +190,7 @@ export function SellerSignup(){
                 onChange={(e)=>{setSellerPass(e.target.value)}}/>
 
                 <label className="formElement">Confirm your password :&emsp;</label>
-                <input className="formElement" />
+                <input className="formElement" type="password" onChange={(e)=>{setConfPass(e.target.value)}}/>
 
                 <button type="submit" className="formElement">Submit</button>
